requirements.txt
biopython==1.85
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
numpy==2.3.3
requests==2.32.5
urllib3==2.5.0
